+++ CHANGELOG.md
@@ -12,6 +12,11 @@ Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Pa
 - Gradle-Files überarbeiten (Deprecated-Aufrufe raus)
 - auf Gradle 9 und Kotlin 2.3 aktualisieren
 
+## [Unreleased]
+### Changed
+- Funktionalität von Name.toString() nach Name.toShortString() verschoben
+
+
 ## [6.6.6]
 ### Changed
 - Reihenfolge der Ausgabe in Name.toString() geändert (zuerst Vorname)